feat: add merge function to tip containers
This may be useful if your plugin also loads tips using the TipLoader
